requirements.txt
# Minimal runtime deps (pip side). Heavy binaries (torch/vision/faiss) via conda.

# Core
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
python-dotenv
bcrypt

# Embeddings & Similarity
sentence-transformers
tqdm
numpy
Pillow
faiss-cpu  # Prefer conda for macOS

# OCR
# torch/torchvision should be installed via conda (see environment.yml)
python-doctr
opencv-python-headless
rapidfuzz

# datasets
kaggle
python-dateutil

# paddle paddle
paddlepaddle-gpu
paddleocr